+++ packages/openai/src/chat/openai-chat-language-model.ts
@@ -11,6 +11,7 @@ import {
 } from '@ai-sdk/provider';
 import {
   FetchFunction,
+  InferFromLazyValidator,
   ParseResult,
   combineHeaders,
   createEventSourceResponseHandler,
@@ -20,14 +21,14 @@ import {
   parseProviderOptions,
   postJsonToApi,
 } from '@ai-sdk/provider-utils';
-import { z } from 'zod/v4';
-import {
-  openaiErrorDataSchema,
-  openaiFailedResponseHandler,
-} from '../openai-error';
+import { openaiFailedResponseHandler } from '../openai-error';
 import { convertToOpenAIChatMessages } from './convert-to-openai-chat-messages';
 import { getResponseMetadata } from './get-response-metadata';
 import { mapOpenAIFinishReason } from './map-openai-finish-reason';
+import {
+  openaiChatChunkSchema,
+  openaiChatResponseSchema,
+} from './openai-chat-api';
 import {
   OpenAIChatModelId,
   openaiChatLanguageModelOptions,
@@ -451,7 +452,7 @@ export class OpenAIChatLanguageModel implements LanguageModelV3 {
     return {
       stream: response.pipeThrough(
         new TransformStream<
-          ParseResult<z.infer<typeof openaiChatChunkSchema>>,
+          ParseResult<InferFromLazyValidator<typeof openaiChatChunkSchema>>,
           LanguageModelV3StreamPart
         >({
           start(controller) {
@@ -696,152 +697,6 @@ export class OpenAIChatLanguageModel implements LanguageModelV3 {
     };
   }
 }
-
-const openaiTokenUsageSchema = z
-  .object({
-    prompt_tokens: z.number().nullish(),
-    completion_tokens: z.number().nullish(),
-    total_tokens: z.number().nullish(),
-    prompt_tokens_details: z
-      .object({
-        cached_tokens: z.number().nullish(),
-      })
-      .nullish(),
-    completion_tokens_details: z
-      .object({
-        reasoning_tokens: z.number().nullish(),
-        accepted_prediction_tokens: z.number().nullish(),
-        rejected_prediction_tokens: z.number().nullish(),
-      })
-      .nullish(),
-  })
-  .nullish();
-
-// limited version of the schema, focussed on what is needed for the implementation
-// this approach limits breakages when the API changes and increases efficiency
-const openaiChatResponseSchema = z.object({
-  id: z.string().nullish(),
-  created: z.number().nullish(),
-  model: z.string().nullish(),
-  choices: z.array(
-    z.object({
-      message: z.object({
-        role: z.literal('assistant').nullish(),
-        content: z.string().nullish(),
-        tool_calls: z
-          .array(
-            z.object({
-              id: z.string().nullish(),
-              type: z.literal('function'),
-              function: z.object({
-                name: z.string(),
-                arguments: z.string(),
-              }),
-            }),
-          )
-          .nullish(),
-        annotations: z
-          .array(
-            z.object({
-              type: z.literal('url_citation'),
-              start_index: z.number(),
-              end_index: z.number(),
-              url: z.string(),
-              title: z.string(),
-            }),
-          )
-          .nullish(),
-      }),
-      index: z.number(),
-      logprobs: z
-        .object({
-          content: z
-            .array(
-              z.object({
-                token: z.string(),
-                logprob: z.number(),
-                top_logprobs: z.array(
-                  z.object({
-                    token: z.string(),
-                    logprob: z.number(),
-                  }),
-                ),
-              }),
-            )
-            .nullish(),
-        })
-        .nullish(),
-      finish_reason: z.string().nullish(),
-    }),
-  ),
-  usage: openaiTokenUsageSchema,
-});
-
-// limited version of the schema, focussed on what is needed for the implementation
-// this approach limits breakages when the API changes and increases efficiency
-const openaiChatChunkSchema = z.union([
-  z.object({
-    id: z.string().nullish(),
-    created: z.number().nullish(),
-    model: z.string().nullish(),
-    choices: z.array(
-      z.object({
-        delta: z
-          .object({
-            role: z.enum(['assistant']).nullish(),
-            content: z.string().nullish(),
-            tool_calls: z
-              .array(
-                z.object({
-                  index: z.number(),
-                  id: z.string().nullish(),
-                  type: z.literal('function').nullish(),
-                  function: z.object({
-                    name: z.string().nullish(),
-                    arguments: z.string().nullish(),
-                  }),
-                }),
-              )
-              .nullish(),
-            annotations: z
-              .array(
-                z.object({
-                  type: z.literal('url_citation'),
-                  start_index: z.number(),
-                  end_index: z.number(),
-                  url: z.string(),
-                  title: z.string(),
-                }),
-              )
-              .nullish(),
-          })
-          .nullish(),
-        logprobs: z
-          .object({
-            content: z
-              .array(
-                z.object({
-                  token: z.string(),
-                  logprob: z.number(),
-                  top_logprobs: z.array(
-                    z.object({
-                      token: z.string(),
-                      logprob: z.number(),
-                    }),
-                  ),
-                }),
-              )
-              .nullish(),
-          })
-          .nullish(),
-        finish_reason: z.string().nullish(),
-        index: z.number(),
-      }),
-    ),
-    usage: openaiTokenUsageSchema,
-  }),
-  openaiErrorDataSchema,
-]);
 
 function isReasoningModel(modelId: string) {
   return (
